requirements.txt
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tensorboard>=2.13.0
wandb>=0.15.0
pydantic>=2.0.0
PyYAML>=6.0
tqdm>=4.65.0
hydra-core>=1.3.0
optuna>=3.2.0
joblib>=1.3.0
